backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==24.1.1
idna==3.11
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
ruff==0.14.13
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
